Move onRouteChange to app store and track route state in store
